requirements.txt
# Claude Agent SDK - Required for the harness itself
claude-code-sdk>=0.0.25

# Environment variable management
python-dotenv>=1.0.0

# Enhanced console output with animations
rich>=13.7.0

# ============================================================
# Generated Project Dependencies (installed by the agent)
# ============================================================
# The autonomous agent will install these in generated projects.
# Listed here for reference only - not needed for the harness.
#
# Core Framework:
# fastapi>=0.109.0
# uvicorn[standard]>=0.27.0
#
# Database:
# sqlalchemy[asyncio]>=2.0.25
# asyncpg>=0.29.0
# alembic>=1.13.0
#
# Validation & Settings:
# pydantic>=2.5.0
# pydantic-settings>=2.1.0
#
# Testing:
# pytest>=7.4.0
# pytest-asyncio>=0.23.0
# httpx>=0.26.0
# pytest-cov>=4.1.0
#
# Authentication (Future):
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.4
#
# Forms:
# python-multipart>=0.0.6
#
# Development Tools (Optional):
# ruff>=0.1.0
# black>=23.12.0
# mypy>=1.8.0
